New class Controller for FIM logic, New static class LogHelper for writing event logs, - getting list of remote connections, SQLiteHelper class is now dynamic with a dispose method
