+++ Project Source Code/WinFIM.NET Service/LogHelper.cs
@@ -1,0 +1,72 @@
+﻿using Serilog;
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Runtime.InteropServices;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace WinFIM.NET_Service
+{
+    internal static class LogHelper
+    {
+        [DllImport("kernel32.dll", SetLastError = true)]
+        private static extern int Wow64DisableWow64FsRedirection(ref IntPtr ptr);
+
+        private static readonly EventLog EventLog1 = new EventLog();
+
+        [DllImport("kernel32.dll", SetLastError = true)]
+        private static extern int Wow64EnableWow64FsRedirection(ref IntPtr ptr);
+
+        internal static void Initialize()
+        {
+            if (!EventLog.SourceExists("WinFIM.NET"))
+            {
+                EventLog.CreateEventSource("WinFIM.NET", "WinFIM.NET");
+            }
+        }
+
+        internal static void WriteEventLog(string message, EventLogEntryType eventType, int eventId)
+        {
+            EventLog1.Source = "WinFIM.NET";
+            EventLog1.Log = "WinFIM.NET";
+            EventLog1.WriteEntry(message, eventType, eventId);
+        }
+
+        internal static string GetRemoteConnections()
+        {
+            string output = "ERROR in running CMD \"query user\"";
+            try
+            {
+                using (Process process = new Process())
+                {
+                    IntPtr val = IntPtr.Zero;
+                    _ = Wow64DisableWow64FsRedirection(ref val);
+                    process.StartInfo.FileName = @"cmd.exe";
+                    process.StartInfo.Arguments = "/c \"@echo off & @for /f \"tokens=1,2,3,4,5\" %A in ('netstat -ano ^| findstr ESTABLISHED ^| findstr /v 127.0.0.1') do (@for /f \"tokens=1,2,5\" %F in ('qprocess \"%E\"') do (@IF NOT %H==IMAGE @echo %A , %B , %C , %D , %E , %F , %G , %H))\"";
+                    process.StartInfo.CreateNoWindow = true;
+                    process.StartInfo.UseShellExecute = false;
+                    process.StartInfo.RedirectStandardOutput = true;
+                    process.Start();
+                    // Synchronously read the standard output of the spawned process. 
+                    output = "Established Remote Connection (snapshot)" + "\n";
+                    output = output + "========================================" + "\n" + "Proto | Local Address | Foreign Address | State | PID | USERNAME | SESSION NAME | IMAGE\n";
+                    output += process.StandardOutput.ReadToEnd();
+                    Log.Verbose(output);
+                    process.WaitForExit();
+                    _ = Wow64EnableWow64FsRedirection(ref val);
+                    return output;
+                }
+
+            }
+            catch (Exception e)
+            {
+                string errorMessage = $"Error in GetRemoteConnections : {e.Message}";
+                Log.Error(errorMessage);
+                return output + "\n" + e.Message;
+            }
+
+        }
+    }
+}
