+++ Project Source Code/WinFIM.NET Service/SQLiteHelper.cs
@@ -5,8 +5,18 @@ using System.IO;
 
 namespace WinFIM.NET_Service
 {
-    internal static class SQLiteHelper
+    internal class SQLiteHelper : IDisposable
     {
+        internal SQLiteConnection Connection { get; }
+
+        private bool _disposed;
+
+        internal SQLiteHelper(string connectionString)
+        {
+            Connection = new SQLiteConnection(connectionString);
+            Connection.Open();
+        }
+
         internal static void EnsureDatabaseExists(string dbFile)
         // Create the database if it doesn't exist
         {
@@ -21,7 +31,7 @@ namespace WinFIM.NET_Service
             }
         }
 
-        internal static void EnsureTablesExist(string connectionString)
+        internal void EnsureTablesExist()
         // Ensure that all required tables exist
         {
             Log.Debug("Creating SQlite table baseline_table if it doesn't exist...");
@@ -34,14 +44,14 @@ namespace WinFIM.NET_Service
                     filehash    TEXT,
                     filetype    TEXT NOT NULL
                 );";
-            ExecuteNonQuery(connectionString, sql);
+            ExecuteNonQuery(sql);
             Log.Debug("Creating SQlite table conf_file_checksum if it doesn't exist...");
             sql = @"
                 CREATE TABLE IF NOT EXISTS conf_file_checksum (
                     pathname    TEXT PRIMARY KEY,
                     filehash    TEXT
                 );";
-            ExecuteNonQuery(connectionString, sql);
+            ExecuteNonQuery(sql);
 
             Log.Debug("Creating SQlite table current_table if it doesn't exist...");
             sql = @"
@@ -53,7 +63,7 @@ namespace WinFIM.NET_Service
                     filehash    TEXT,
                     filetype    TEXT NOT NULL
                 );";
-            ExecuteNonQuery(connectionString, sql);
+            ExecuteNonQuery(sql);
 
             Log.Debug("Creating SQlite table monlist if it doesn't exist...");
             sql = @"
@@ -62,7 +72,7 @@ namespace WinFIM.NET_Service
                     pathexists BOOLEAN  CHECK (pathexists IN (0, 1)) NOT NULL,
                     checktime TEXT NOT NULL
                 );";
-            ExecuteNonQuery(connectionString, sql);
+            ExecuteNonQuery(sql);
 
             Log.Debug("Creating SQlite table version_control if it doesn't exist...");
             sql = @"
@@ -70,72 +80,82 @@ namespace WinFIM.NET_Service
                     version     INT PRIMARY KEY,
                     notes       TEXT NOT NULL
                 );";
-            ExecuteNonQuery(connectionString, sql);
+            ExecuteNonQuery(sql);
 
             Log.Debug("Setting database version...");
             sql = @"
                 INSERT OR REPLACE INTO version_control (version, notes) VALUES (1, 'Initial version');
                 INSERT OR REPLACE INTO version_control (version, notes) VALUES (2, 'changed field filename to pathname and set as primary key;\n added field isbasepath to tables baseline_table, current_table; added tables version_control, monlist_baseline_table, monlist_current_table');
             ";
-            ExecuteNonQuery(connectionString, sql);
+            ExecuteNonQuery(sql);
         }
 
-        internal static void ExecuteNonQuery(string connectionString, string sql)
-        // To test SQL table population
+        internal void ExecuteNonQuery(string sql)
         {
-            using (SQLiteConnection connection = new SQLiteConnection(connectionString))
+            try
             {
-                try
+                using (SQLiteCommand command = new SQLiteCommand(Connection))
                 {
-                    connection.Open();
-                    using (SQLiteCommand command = new SQLiteCommand(connection))
-                    {
-                        Log.Verbose($"Running ExecuteNonQuery {sql}");
-                        command.CommandText = sql;
-                        command.ExecuteNonQuery();
-                    }
-                }
-                catch (Exception e)
-                {
-                    string errorMessage = $"Error running ExecuteNonQuery {sql}";
-                    Log.Error(e, errorMessage);
-                    throw;
-                }
-                finally
-                {
-                    connection.Close();
+                    Log.Verbose($"Running ExecuteNonQuery {sql}");
+                    command.CommandText = sql;
+                    command.ExecuteNonQuery();
                 }
             }
+            catch (Exception e)
+            {
+                string errorMessage = $"Error running ExecuteNonQuery {sql}";
+                Log.Error(e, errorMessage);
+                throw;
+            }
+
         }
 
         // A query that returns the first value in the first row as an object
-        internal static object ExecuteScalar(string connectionString, string sql)
+        internal object ExecuteScalar(string sql)
         {
-            object output = null;
-            using (SQLiteConnection connection = new SQLiteConnection(connectionString))
+            object output;
+            try
             {
-                try
+                using (SQLiteCommand command = new SQLiteCommand(Connection))
                 {
-                    connection.Open();
-                    using (SQLiteCommand command = new SQLiteCommand(connection))
-                    {
-                        Log.Verbose($"Running ExecuteScalarAsync {sql}");
-                        command.CommandText = sql;
-                        output = command.ExecuteScalarAsync().Result;
-                    }
-                }
-                catch (Exception e)
-                {
-                    string errorMessage = $"Error running query {sql}";
-                    Log.Error(e, errorMessage);
-                    throw;
-                }
-                finally
-                {
-                    connection.Close();
+                    Log.Verbose($"Running ExecuteScalarAsync {sql}");
+                    command.CommandText = sql;
+                    output = command.ExecuteScalarAsync().Result;
                 }
             }
+            catch (Exception e)
+            {
+                string errorMessage = $"Error running query {sql}";
+                Log.Error(e, errorMessage);
+                throw;
+            }
+
             return output;
+        }
+
+        public void Dispose()
+        {
+            Dispose(disposing: true);
+            GC.SuppressFinalize(this);
+        }
+
+        protected virtual void Dispose(bool disposing)
+        {
+            // Check to see if Dispose has already been called.
+            if (!this._disposed)
+            {
+                // If disposing equals true, dispose all managed
+                // and unmanaged resources.
+                if (disposing)
+                {
+                    // Dispose managed resources.
+                    Connection.Close();
+                    Connection.Dispose();
+                }
+
+                // Note disposing has been done.
+                _disposed = true;
+            }
         }
     }
 }
